Modify functionality for Drawer View expansion
Rename variables associated with the view that shows hourly forecasts to DrawerView. Modify toggleDrawerViewExpansion function to add or subtract to the constraint of the drawerView by a fixed value depending on if already hidden or not. Give initial constraint of the Drawer View in Storyboard to be 90% less than height. Clean up spacing and removal of unnecessary code.