Modify model of app to be mutable
Change the type of BeachForecast from Class to Struct and change its forecast from var to let along with a new init. Modify retrieveBeachNames to pass an array of Beach within the completion handler. Modify updateForecasts to take an array of Beach and fetch forecast for each one. Make a new object of BeachForecast given the new Beach and Forecast then append to a new list of BeachForecasts. Assign the beachForecasts array to this new array.