Modify Beach Forecast Model Controller by making core objects private vars
Make the arrays that hold the Beach Forecasts, the service managers, and the user location private vars. Add computed variable that returns the count of one of the Beach Forecast Arrays. Add function that returns a Beach Forecast at a given index. Add updateLocation function to tell the location manager to start updating location instead of directly asking the location manager to do so.